+++ src/constants.ts
@@ -6,10 +6,11 @@ import {
   type PaymentFlow,
   type CheckoutStyle,
   type HeadlessUniversalCheckoutOptions,
+  type ApplePayOptions,
 } from '@primer-io/checkout-web';
 import { PaymentMethod } from './enums';
 
-export const SDK_VERSION = '0.6.2';
+export const SDK_VERSION = '0.6.3';
 
 export const DEFAULTS = {
   BASE_URL: 'https://billing.funnelfox.com',
@@ -179,5 +180,17 @@ export const DEFAULT_BUTTONS_OPTIONS: {
     buttonColor: 'black',
     buttonSizeMode: 'fill',
     buttonType: 'pay',
+  },
+};
+
+export const APPLE_PAY_COLLECTING_EMAIL_OPTIONS: Pick<
+  ApplePayOptions,
+  'billingOptions' | 'shippingOptions'
+> = {
+  billingOptions: {
+    requiredBillingContactFields: ['emailAddress'],
+  },
+  shippingOptions: {
+    requiredShippingContactFields: ['emailAddress'],
   },
 };