+++ CHANGELOG.md
@@ -179,3 +179,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - prevent primer payment request after checkout destroy
 - change google pay icon in default skin
 - add Primer types
+
+## [0.6.3] - 2025-02-23
+
+- add apple pay email props by deature flag
+- fix PURCHASE_CANCELLED event